Refactor(trees): Modularize trees.h into isolated data structure headers
