+++ demos/trees/bplus_tree_demo.c
@@ -1,6 +1,6 @@
+#include "bplus_tree.h"
 #include "mwst_utils.h"
 #include "safe_input.h"
-#include "trees.h"
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
