+++ demos/trees/fenwick_tree_demo.c
@@ -1,6 +1,6 @@
 #include "display_header.h"
+#include "fenwick_tree.h"
 #include "safe_input.h"
-#include "trees.h"
 #include <stdio.h>
 #include <stdlib.h>
