+++ demos/trees/tbt_demo.c
@@ -1,5 +1,5 @@
 #include "safe_input.h"
-#include "trees.h"
+#include "tbt.h"
 #include <stdio.h>
 #include <stdlib.h>
