+++ demos/trees/trie_demo.c
@@ -1,5 +1,5 @@
 #include "safe_input.h"
-#include "trees.h"
+#include "trie.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
