+++ features/file_exporter/exporter_trees.c
@@ -2,10 +2,10 @@
 
 bool export_bst(const char* root_dir, const char* dest_dir)
 {
-    return export_file_pair(root_dir, "bst", "trees", dest_dir, NULL, NULL);
+    return export_file_pair(root_dir, "bst", "bst", dest_dir, NULL, NULL);
 }
 
 bool export_avl(const char* root_dir, const char* dest_dir)
 {
-    return export_file_pair(root_dir, "avl", "trees", dest_dir, NULL, NULL);
+    return export_file_pair(root_dir, "avl", "avl", dest_dir, NULL, NULL);
 }
