+++ src/trees/fenwick_tree.c
@@ -1,5 +1,5 @@
+#include "fenwick_tree.h"
 #include "safe_input.h"
-#include "trees.h"
 #include <stdio.h>
 #include <stdlib.h>
