+++ src/trees/red_black_tree.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "red_black_tree.h"
 #include <stdio.h>
 #include <stdlib.h>
