+++ src/trees/splay_tree.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "splay_tree.h"
 #include <stdio.h>
 #include <stdlib.h>
