+++ src/trees/tbt.c
@@ -1,5 +1,5 @@
+#include "tbt.h"
 #include "safe_input.h"
-#include "trees.h"
 #include <stdio.h>
 #include <stdlib.h>
