+++ src/trees/trie.c
@@ -1,5 +1,5 @@
+#include "trie.h"
 #include "safe_input.h"
-#include "trees.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
