+++ tests/trees/test_avl.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "avl.h"
 #include <assert.h>
 #include <stdio.h>
