+++ tests/trees/test_bplus_tree.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "bplus_tree.h"
 #include <assert.h>
 #include <stdbool.h>
 #include <stdio.h>
