+++ tests/trees/test_btree.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "btree.h"
 #include <assert.h>
 #include <stdio.h>
