+++ tests/trees/test_tbt.c
@@ -1,4 +1,4 @@
-#include "trees.h"
+#include "tbt.h"
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
